podcast_production/requirements.txt
# AI Podcast Production System - LangGraph Production
# Core dependencies for the podcast production workflow
# Version: 2.0.0
# Date: September 2025

# Core workflow orchestration
langgraph>=0.2.5
langgraph-checkpoint>=1.0.2
langgraph-checkpoint-postgres>=1.0.0  # For distributed state

# Observability and monitoring
langfuse>=2.1.0
langsmith>=0.1.5  # APM integration
langchain-core>=0.1.5  # For RunnablePassthrough and callbacks

# HTTP client for API calls
httpx>=0.27.0
aiohttp>=3.9.0  # Alternative async client

# Retry and resilience
tenacity>=8.2.0
circuit-breaker>=1.4.0  # Circuit breaker pattern

# Configuration management
pyyaml>=6.0.1
pydantic>=2.5.0
python-dotenv>=1.0.0

# Type checking and validation
typing-extensions>=4.9.0

# Development and testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0  # Coverage reporting
pytest-mock>=3.12.0  # Mock testing
locust>=2.17.0  # Load testing

# API Provider integrations (REQUIRED)
openai>=1.12.0  # OpenAI GPT-4 integration
anthropic>=0.28.0  # Anthropic Claude integration
google-generativeai>=0.3.5  # Google Gemini integration

# Audio processing (REQUIRED)
elevenlabs>=0.3.0  # ElevenLabs TTS
pydub>=0.25.1  # Audio manipulation
librosa>=0.10.1  # Audio analysis

# Vector databases and embeddings
chromadb>=0.4.22  # Vector store for brand consistency
sentence-transformers>=2.3.0  # Embeddings for brand analysis

# Monitoring and analytics
datadog>=0.47.0  # APM alternative
prometheus-client>=0.19.0  # Metrics export
redis>=5.0.0  # Caching and state

# Security
cryptography>=41.0.7  # Encryption
python-jose>=3.3.0  # JWT handling
hvac>=2.0.0  # HashiCorp Vault client

# Utilities
msgpack>=1.0.7  # Efficient serialization
ujson>=5.9.0  # Fast JSON
asyncio-throttle>=1.0.0  # Rate limiting